[magenta] Replace use of sys/types.h with stdint.h
sys/types.h is a POSIX header files and is not a part of C
standard. There are number of defines in sys/types.h which are
guarded by either _GNU_SOURCE or _BSD_SOURCE. We should be using
stdint.h and stddef.h instead which are C standard equivalents.

+++ system/ulib/mxio/elf.c
@@ -279,7 +279,7 @@ mx_status_t elf_load(elf_handle_t* handle) {
         handle->eheader.e_phnum * sizeof(elf_phdr_t);
 
     LTRACEF("program headers:\n");
-    for (uint i = 0; i < handle->eheader.e_phnum; i++) {
+    for (size_t i = 0; i < handle->eheader.e_phnum; i++) {
         // parse the program headers
         elf_phdr_t* pheader = &handle->pheaders[i];
